requirements.txt
# pip install --upgrade --force-reinstall -r requirements.txt

invoke==1.3.0
ruamel.yaml==0.16.5
Jinja2==2.10.3
pytest==5.2.2
faker==2.0.3
black==19.10b0
pre-commit==1.20.0
